requirements.txt
uvicorn
fastapi
firebase_admin
pydantic
pymongo
python-multipart
PyPDF2
python-jose
passlib
bcrypt
google-auth 
google-auth-oauthlib
itsdangerous
python-dotenv
wraps
pyjwt==1.7.1


pyrebase4
requests==2.29.0

stripe
